refactor: Move mock setup to be shared
