+++ src/entity-search/__fixtures__/company-search-no-results.json
