+++ src/entity-search/__fixtures__/company-search-some-other-company.json
