+++ src/entity-search/__fixtures__/company-search.json
